requirements.txt
imagekitio==4.2.0
pandas==2.3.2
Pillow==11.3.0
pymongo==4.15.1
Requests==2.32.5
streamlit==1.48.0
